Add validations to DTO, Try Catch to Repository class, validation for event Status before post and StatusCodes to Controllers.

+++ LMCEvents.Core/Interfaces/IEventReservationRepository.cs
@@ -14,7 +14,7 @@ namespace LMCEvents.Core.Interfaces
 
         bool InsertBooking(EventReservation booking);
 
-        bool UpdateBooking(long idBooking, int newQuantity);
+        bool UpdateBooking(long idBooking, long newQuantity);
 
         bool DeleteBooking(long idBooking);
